requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
altair==4.2.0; python_version >= '3.7'
argon2-cffi-bindings==21.2.0; python_version >= '3.6'
argon2-cffi==21.3.0; python_version >= '3.6'
asttokens==2.0.5
attrs==21.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
backcall==0.2.0
beautifulsoup4==4.11.1; python_version >= '3.6'
bleach==5.0.0; python_version >= '3.7'
blinker==1.4
cachetools==5.2.0; python_version ~= '3.7'
certifi==2022.5.18.1; python_version >= '3.6'
cffi==1.15.0
charset-normalizer==2.0.12; python_version >= '3'
click==8.1.3; python_version >= '3.7'
commonmark==0.9.1
cycler==0.11.0; python_version >= '3.6'
debugpy==1.6.0; python_version >= '3.7'
decorator==5.1.1; python_version >= '3.5'
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
entrypoints==0.4; python_version >= '3.6'
et-xmlfile==1.1.0; python_version >= '3.6'
executing==0.8.3
fastjsonschema==2.15.3
fonttools==4.33.3; python_version >= '3.7'
gitdb==4.0.9; python_version >= '3.6'
gitpython==3.1.27; python_version >= '3.7'
idna==3.3; python_version >= '3'
importlib-metadata==4.11.4; python_version >= '3.7'
ipykernel==6.13.1; python_version >= '3.4'
ipython-genutils==0.2.0
ipython==8.4.0; python_version >= '3.8'
ipywidgets==7.7.0
jedi==0.18.1; python_version >= '3.6'
jinja2==3.1.2; python_version >= '3.7'
jsonschema==4.6.0; python_version >= '3.7'
jupyter-client==7.3.4; python_version >= '3.7'
jupyter-core==4.10.0; python_version >= '3.7'
jupyterlab-pygments==0.2.2; python_version >= '3.7'
jupyterlab-widgets==1.1.0; python_version >= '3.6'
kiwisolver==1.4.2; python_version >= '3.7'
markupsafe==2.1.1; python_version >= '3.7'
matplotlib-inline==0.1.3; python_version >= '3.5'
matplotlib==3.5.2
mistune==0.8.4
nbclient==0.6.4; python_version >= '3.7'
nbconvert==6.5.0; python_version >= '3.7'
nbformat==5.4.0; python_version >= '3.7'
nest-asyncio==1.5.5; python_version >= '3.5'
notebook==6.4.12; python_version >= '3.7'
numpy==1.22.4
openpyxl==3.0.10
packaging==21.3; python_version >= '3.6'
pandas==1.3
pandocfilters==1.5.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
parso==0.8.3; python_version >= '3.6'
patsy==0.5.2
pexpect==4.8.0; sys_platform != 'win32'
pickleshare==0.7.5
pillow==9.1.1; python_version >= '3.7'
prometheus-client==0.14.1; python_version >= '3.6'
prompt-toolkit==3.0.29; python_full_version >= '3.6.2'
protobuf==3.20.1; python_version >= '3.7'
psutil==5.9.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==8.0.0; python_version >= '3.7'
pycparser==2.21
pydeck==0.7.1; python_version >= '3.7'
pygments==2.12.0; python_version >= '3.6'
pympler==1.0.1; python_version >= '3.6'
pyparsing==3.0.9; python_full_version >= '3.6.8'
pyrsistent==0.18.1; python_version >= '3.7'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz-deprecation-shim==0.1.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pytz==2022.1
pyzmq==23.1.0; python_version >= '3.6'
requests==2.27.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
rich==12.4.4; python_full_version >= '3.6.3' and python_full_version < '4.0.0'
scipy==1.8.1
seaborn==0.11.2
semver==2.13.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
send2trash==1.8.0
setuptools==62.3.3; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
smmap==5.0.0; python_version >= '3.6'
soupsieve==2.3.2.post1; python_version >= '3.6'
stack-data==0.2.0
statsmodels==0.13.2
streamlit==1.10.0
terminado==0.15.0; python_version >= '3.7'
tinycss2==1.1.1; python_version >= '3.6'
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
toolz==0.11.2; python_version >= '3.5'
tornado==6.1; python_version >= '3.5'
traitlets==5.2.2.post1; python_version >= '3.7'
typing-extensions==4.2.0; python_version >= '3.7'
tzdata==2022.1; python_version >= '3.6'
tzlocal==4.2; python_version >= '3.6'
urllib3==1.26.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_full_version < '4.0.0'
validators==0.20.0; python_version >= '3.4'
watchdog==2.1.8; platform_system != 'Darwin'
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.0
zipp==3.8.0; python_version >= '3.7'
